split out street icons
