+++ src/editor/icons/icons.jsx
@@ -127,66 +127,6 @@ export const GeospatialIconWhite = () => (
   </svg>
 );
 
-export const ArrowLeftHookIcon = () => (
-  <svg
-    width="26"
-    height="26"
-    viewBox="0 0 26 26"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M10.3245 21.1578C10.6418 20.8405 10.6418 20.3261 10.3245 20.0088L7.3782 17.0625H17.3333C18.6981 17.0625 20.0069 16.5203 20.972 15.5553C21.937 14.5903 22.4792 13.2814 22.4792 11.9166C22.4792 10.5519 21.937 9.24302 20.972 8.27799C20.0069 7.31296 18.6981 6.77081 17.3333 6.77081H16.25C15.8013 6.77081 15.4375 7.13458 15.4375 7.58331C15.4375 8.03204 15.8013 8.39581 16.25 8.39581H17.3333C18.2671 8.39581 19.1626 8.76676 19.8229 9.42704C20.4832 10.0873 20.8542 10.9829 20.8542 11.9166C20.8542 12.8504 20.4832 13.746 19.8229 14.4063C19.1626 15.0665 18.2671 15.4375 17.3333 15.4375H7.3782L10.3245 12.4912C10.6418 12.1739 10.6418 11.6594 10.3245 11.3421C10.0072 11.0248 9.49277 11.0248 9.17547 11.3421L4.84246 15.6751C4.84235 15.6752 4.84224 15.6753 4.84213 15.6755C4.84156 15.676 4.84098 15.6766 4.84041 15.6772C4.76337 15.7547 4.70517 15.8438 4.66581 15.939C4.62608 16.0348 4.60416 16.1398 4.60416 16.25C4.60416 16.4735 4.69439 16.6759 4.84041 16.8228C4.84098 16.8234 4.84156 16.8239 4.84213 16.8245C4.84224 16.8246 4.84235 16.8247 4.84246 16.8248L9.17547 21.1578C9.49277 21.4751 10.0072 21.4751 10.3245 21.1578Z"
-      fill="white"
-    />
-  </svg>
-);
-
-export const ManualIcon = () => (
-  <svg
-    width="25"
-    height="30"
-    viewBox="0 0 25 30"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M5.15243 7.90909H8.56578L12.1709 16.7045H12.3243L15.9294 7.90909H19.3428V21H16.6581V12.4794H16.5495L13.1617 20.9361H11.3335L7.94576 12.4474H7.83709V21H5.15243V7.90909Z"
-      fill="#EA4335"
-    />
-    <path
-      d="M12.2572 28.8995L0.5 20.9199V8.37301L12.2571 0.595916L24.5 8.37889V20.914L12.2572 28.8995Z"
-      fill="#EA4335"
-      fillOpacity="0.23"
-      stroke="#EA4335"
-    />
-  </svg>
-);
-
-export const AutoIcon = () => (
-  <svg
-    width="27"
-    height="26"
-    viewBox="0 0 27 26"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <circle
-      cx="13.5"
-      cy="13"
-      r="12.5"
-      fill="#0EAF00"
-      fillOpacity="0.17"
-      stroke="#0EAF00"
-    />
-    <path
-      d="M10.1618 19H7.19585L11.715 5.90909H15.2818L19.7946 19H16.8287L13.5495 8.90057H13.4473L10.1618 19ZM9.97638 13.8544H16.9821V16.0149H9.97638V13.8544Z"
-      fill="#0EAF00"
-    />
-  </svg>
-);
 export const Camera32Icon = () => (
   <svg
     width="32"
@@ -958,102 +898,5 @@ export const SignInMicrosoftIconSVG = () => (
     <rect x="13" y="21" width="9" height="9" fill="#00a4ef" />
     <rect x="23" y="11" width="9" height="9" fill="#7fba00" />
     <rect x="23" y="21" width="9" height="9" fill="#ffb900" />
-  </svg>
-);
-
-export const ManagedStreetIcon = () => (
-  <svg
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M17 1L17 22"
-      stroke="white"
-      strokeLinecap="round"
-      strokeDasharray="3 4"
-    />
-    <path
-      d="M6 1L6 22"
-      stroke="white"
-      strokeLinecap="round"
-      strokeDasharray="3 4"
-    />
-    <path
-      d="M12 1L12 22"
-      stroke="white"
-      strokeWidth="0.5"
-      strokeLinecap="round"
-    />
-    <path
-      d="M11 1L11 22"
-      stroke="white"
-      strokeWidth="0.5"
-      strokeLinecap="round"
-    />
-    <rect
-      x="0.75"
-      y="0.75"
-      width="21.5"
-      height="21.5"
-      rx="1.25"
-      stroke="#00FF37"
-      strokeWidth="1.5"
-    />
-  </svg>
-);
-
-export const SegmentIcon = () => (
-  <svg
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M17 1L17 22"
-      stroke="white"
-      strokeLinecap="round"
-      strokeDasharray="3 4"
-    />
-    <path
-      d="M6 1L6 22"
-      stroke="white"
-      strokeLinecap="round"
-      strokeDasharray="3 4"
-    />
-    <path
-      d="M12 1L12 22"
-      stroke="white"
-      strokeWidth="0.5"
-      strokeLinecap="round"
-    />
-    <path
-      d="M11 1L11 22"
-      stroke="white"
-      strokeWidth="0.5"
-      strokeLinecap="round"
-    />
-    <rect
-      x="0.75"
-      y="0.75"
-      width="21.5"
-      height="21.5"
-      rx="1.25"
-      stroke="#929292"
-      strokeWidth="1.5"
-    />
-    <rect
-      x="9.75"
-      y="0.75"
-      width="6.5"
-      height="21.5"
-      rx="1.25"
-      stroke="#00FF37"
-      strokeWidth="1.5"
-    />
   </svg>
 );
