+++ src/editor/icons/index.js
@@ -1,1 +1,2 @@
 export * from './icons.jsx';
+export * from './street-icons.jsx';
